refactor api interface structure
